requirements.txt
environs==14.1.1
google-cloud-dialogflow==2.41.1
python-telegram-bot==13.15
requests==2.32.3
vk-api==11.9.9
